refactor: pull out setProperty method
