+++ lambdas/authorize.js
@@ -3,6 +3,7 @@
 const https = require('https')
 const querystring = require('querystring')
 const { SSM, CognitoIdentity, CognitoSync } = require('aws-sdk')
+const { setProperty } = require('common')
 
 const ssm = new SSM()
 
@@ -145,21 +146,4 @@ exports.handler = async (event) => {
     req.write(postData)
     req.end()
   })
-}
-
-/**
- * Recursively set object properties on obj. Path is a slash separate path.
- */
-const setProperty = (obj, path, value) => {
-  const s = path.split('/', 2)
-  const k = s[0]
-  if (s.length === 2) {
-    if (!obj[k]) {
-      obj[k] = {}
-    }
-    obj[k] = setProperty(obj[k], s[1], value)
-  } else {
-    obj[k] = value
-  }
-  return obj
 }
